sports_booking/requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
django==4.2.27
djangorestframework==3.15.2
pg8000==1.31.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
scramp==1.4.6
six==1.17.0
sqlparse==0.5.4
typing-extensions==4.13.2
tzdata==2025.2
